+++ src/REvent.tsx
@@ -37,6 +37,19 @@ export class RlayersBase<P, S> extends React.PureComponent<P, S> {
         }
     }
 
+    // Used when replacing a source
+    attachOldEventHandlers(newSource: BaseObject): void {
+        // No events have been attached yet
+        if (!this.handlers) return;
+        const events = Object.keys(this.props).filter((p) => p.startsWith('on'));
+        for (const e of events) {
+            if (this.props[e]) {
+                debug('reinstalling existing handler', this, e, this.props[e]);
+                newSource.on(this.olEventName(e), this.handlers[e]);
+            }
+        }
+    }
+
     refresh(prevProps?: P): void {
         this.attachEventHandlers();
     }