+++ src/layer/RLayerTile.tsx
@@ -56,10 +56,10 @@ export default class RLayerTile extends RLayerRaster<RLayerTileProps> {
 
     refresh(prevProps?: RLayerTileProps): void {
         super.refresh(prevProps);
-        if (prevProps?.tileGrid !== this.props.tileGrid) this.createSource();
-        if (this.props.url && prevProps?.url !== this.props.url) {
-            this.source.setUrl(this.props.url);
-            this.source.refresh();
+        if (prevProps?.tileGrid !== this.props.tileGrid || prevProps?.url !== this.props.url) {
+            this.createSource();
+            this.ol.setSource(this.source);
+            this.attachOldEventHandlers(this.source);
         }
     }
 }