+++ CHANGELOG.md
@@ -5,7 +5,7 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
-### [1.4.8] WIP
+### [1.4.8] 2023-01-08
 
 -   Support additional `<RMap>` properties, including disabling of the rotation
 -   Fix [#109], interactions do not support JSX styles